fix build errors

+++ api/src/main/kotlin/dev/zemco/schemalens/modeling/api/dtos/DataModelRelationshipAttributeDtos.kt
@@ -3,5 +3,5 @@ package dev.zemco.schemalens.modeling.api.dtos
 data class DataModelRelationshipAttributeDto(
     val referencedAttributeId: Long,
     val name: String,
-    val position: Int
+    val position: Short
 )
